Add Pipeline class, modify ParamsProvider/PipelineComponent inheritance logic
